Lay out APIs for linking / resolving references in both code and data
Labels and the reference resolver are being factored out of the x86
assembler, and into a more generic "linker". That way, labels can also
be used for within the data segment, and the code and data can reference
each other without having to _a priori_ know the location of the label.

It is not functionally complete yet, but the public APIs are mostly
finished.

Previous versions of the compiler used to calculate addresses before
referencing them. Labels were implemented for code when the assembler
was made, but data addresses were still pre-calculated, which was
getting pretty cumbersome as more static data structures were added.